unpack DL101 course nbs
The course notebooks are named with relation to the Bootcamp structure.

Course notebooks are adopted from the most recent fastai notebooks that were used for DL101 Cohort #4.

Each course nb contains the description and the link to the Weekly Deliverables (were shared each week with students to be prepared for the Mentor Session)